sales and purchase kısmı tamamen tamamlandı ürünler stoktan düşüyor ve ekleniyor total price bulunuyor

+++ MyApp.Core/Entities/Purchase.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -12,7 +13,7 @@ namespace MyAppCore.Entities
         public int StockId { get; set; }
         public Stock Stock { get; set; }
         public int Quantity { get; set; }
-        public decimal TotalCost { get; set; }
+        public int TotalCost { get; set; }
         public DateTime PurchaseDate { get; set; }
     }
